pyproject.toml
[tool.poetry]
name = "veterinary-assistant"
version = "0.1.0"
description = "A veterinary medical assistant application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "3.0.0"
google-generativeai = "0.3.2"
python-dotenv = "1.0.0"
pymongo = "4.6.1"
Pillow = "10.1.0"
SpeechRecognition = "3.10.0"
googletrans = "4.0.0-rc1"
pyttsx3 = "2.90"
requests = "2.31.0"
pandas = "2.1.4"
plotly = "5.18.0"
python-weather = "1.1.1"
opencv-python-headless = "4.8.1.78"
bcrypt = "4.1.2"
gunicorn = "21.2.0"
Werkzeug = "3.0.1"
click = "8.1.7"
itsdangerous = "2.1.2"
Jinja2 = "3.1.2"
MarkupSafe = "2.1.3"
reportlab = "4.0.8"
joblib = "1.3.2"
scikit-learn = "1.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
